+++ README.md
@@ -26,9 +26,11 @@ This is what we are currently working on:
   - Bug fix for 3D6 roll icon (at bottom of screen)
   - Bug fix for costs, counts, weights that contain commas ","
   - Bug fix for 17&18 always being a failure (even with very high skill)
-  - Updated GCA export to handle & in ad/disad/quirk/perk name
+  - Updated GCA export to handle '&' in ad/disad/quirk/perk name
   - Bug fix for drag and drop reordering in lists
   - For GMs, display LastActor name above Modifier Bucket
+  - Allow full damage types, e.g. 'crush', 'cutting', etc.
+  - Bug fix for NPC bodyplan changes
 
 0.8.14 - 2/11/2021
