+++ lib/parselink.js
@@ -1,7 +1,7 @@
 'use strict'
 
 import { GURPS } from "../module/gurps.js";
-import { woundModifiers } from '../module/damage/damage-tables.js'
+import { woundModifiers, damageTypeMap } from '../module/damage/damage-tables.js'
 import { d6ify } from './utilities.js'
 
 
@@ -44,7 +44,8 @@ export function parseForDamage(str) {
   if (!!a) {
     let D = a[DMG_INDEX_D] || ""  // Can now support non-variable damage '2 cut' or '2x3(1) imp'
     const damageType = !!a[DMG_INDEX_TYPE] ? a[DMG_INDEX_TYPE].trim() : ""
-    const woundingModifier = woundModifiers[damageType.toLowerCase()]
+    const dmap = damageTypeMap[damageType.toLowerCase()]
+    const woundingModifier = woundModifiers[dmap]
     const adds = a[DMG_INDEX_ADDS] || ""
     let multiplier = a[DMG_INDEX_MULTIPLIER] || ""
     if (!!multiplier && 'Xx'.includes(multiplier[0])) multiplier = '*' + multiplier.substr(1) // Must convert to '*' for Foundry.
@@ -84,7 +85,8 @@ export function parseForDamage(str) {
   if (!!a) {
     const basic = a[DMG_INDEX_BASICDAMAGE]
     const damageType = !!a[DMG_INDEX_TYPE] ? a[DMG_INDEX_TYPE].trim() : ""
-    const woundingModifier = woundModifiers[damageType.toLowerCase()]
+    const dmap = damageTypeMap[damageType.toLowerCase()]
+    const woundingModifier = woundModifiers[dmap]
     const adds = a[DMG_INDEX_ADDS] || ""
     let multiplier = a[DMG_INDEX_MULTIPLIER] || ""
     if (!!multiplier && 'Xx'.includes(multiplier[0])) multiplier = '*' + multiplier.substr(1) // Must convert to '*' for Foundry.
